mysql dockerize
Created the dockerfile.mysql to create the image database using docker.

The MySQL root password is set to "12345," the database is named "CRS," and a MySQL user with the name "root". The init.sql script is copied to the /docker-entrypoint-initdb.d/ directory, ensuring that it will be executed when the container starts.

dockerfile.mysql:
FROM mysql: Specifies the base image as the official MySQL Docker image with the latest version

ENV: Sets environment variables for MySQL. You can customize these variables based on your requirements. In this example, we set the root password, create a database, and create a user with a password.

COPY init.sql /docker-entrypoint-initdb.d/: Copies an initialization SQL script (assuming you have one) into the /docker-entrypoint-initdb.d/ directory. The SQL script will be executed when the container starts.

EXPOSE 3306: Informs Docker that the container will listen on port 3306.

CMD ["mysqld"]: Specifies the command to run when the container starts. In this case, it starts the MySQL daemon.

init.sql: script for creating tables if not exist.

To create the image:
docker build -t database . -f ~/dockerfile.mysql

To create the container:
docker run -d -p 3306:3306 --name dbcontainer database